# Make Notifications::Fanout faster after changing subscriptions